requirements.txt
# Fikiri Solutions - Production Requirements
# Lightweight, production-ready dependencies

# Core Google API
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# AI Services
openai>=1.0.0

# Microsoft Graph Integration
msal>=1.24.0
requests>=2.31.0

# Twilio (SMS / Verify)
twilio>=9.0.0

# JWT for Apple iCloud Integration
PyJWT>=2.9.0

# Flask JWT Extension
flask-jwt-extended>=4.5.0

# Document Processing
beautifulsoup4>=4.0.0
html5lib>=1.1

# Environment (for .env loading in app and contract tests)
python-dotenv>=1.0.0

# Web Framework
flask>=3.0.0,<4.0.0
flask-socketio>=5.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
flask-restx>=1.2.0
gunicorn>=20.0.0
marshmallow>=3.20.0  # Schema validation for API endpoints

# Monitoring and Alerting
sentry-sdk[flask]>=2.35.0
schedule>=1.2.0
boto3>=1.34.0

# Workflow Automation
fastapi>=0.100.0
uvicorn>=0.20.0
pyyaml>=6.0
click>=8.0

# Strategic Heavy Dependencies (Optional - Use Feature Flags)
# Uncomment only what you need for enhanced features:
# scikit-learn>=1.0.0                    # For enhanced ML scoring
# Transformers removed for lightweight operation
# PyTorch removed for lightweight operation
# ML dependencies removed for lightweight operation

# NumPy compatibility - pin to 1.x to avoid conflicts with scikit-learn
numpy<2.0

# Database and Caching
SQLAlchemy>=2.0.0
redis>=4.0.0
psutil>=5.9.0

# Vector Database (Optional - only if using Pinecone)
# pinecone>=5.0.0  # Uncomment if using Pinecone (package renamed from pinecone-client)

# Security and Encryption
cryptography>=41.0.0

# Development and Testing Tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
aiohttp>=3.9.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0
bandit>=1.7.0
isort>=5.12.0



# Date and timezone handling
pytz>=2023.3
python-dateutil>=2.8.2
# Force rebuild Fri Oct  3 13:22:06 EDT 2025
